+++ dca/tl_anystores.php
@@ -659,21 +659,26 @@ class tl_anystores extends Backend
 
     /**
      * Displays a little static Google Map with position of the address
+     * @todo Make dynamic and change coordinates on click
+     *
      * @param DataContainer
      * @return string
      */
     public function showMap(DataContainer $dc)
     {
-
-        $sCoords = sprintf("%s,%s",
+        $strCoords = sprintf("%s,%s",
                 $dc->activeRecord->latitude,
                 $dc->activeRecord->longitude
         );
 
-        return '<div style="float: right; height: 139px; margin-right: 23px; overflow: hidden; width: 320px;">'
-                .'<h3><label>'.$GLOBALS['TL_LANG']['tl_anystores']['map'][0].'</label></h3> '
-                .'<img style="margin-top: 1px;" src="http://maps.google.com/maps/api/staticmap?center='.$sCoords.'&zoom=16&size=320x139&maptype=roadmap&markers=color:red|label:|'.$sCoords.'&sensor=false" />'
-                .'</div>';
+        return sprintf(
+            '<div style="float: right; height: 139px; margin-right: 23px; overflow: hidden; width: 320px;">
+                <h3><label>%1$s</label></h3>
+                <img style="margin-top:1px;" src="http://maps.google.com/maps/api/staticmap?center=%2$s&zoom=16&size=320x139&maptype=roadmap&markers=color:red|label:|%2$s">
+            </div>',
+            $GLOBALS['TL_LANG']['tl_anystores']['map'][0],
+            $strCoords
+        );
     }